v4 권한 처리 완료

+++ src/main/java/shop/mtcoding/blog/board/BoardController.java
@@ -7,6 +7,9 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import shop.mtcoding.blog._core.errors.exception.Exception401;
+import shop.mtcoding.blog._core.errors.exception.Exception403;
+import shop.mtcoding.blog._core.errors.exception.Exception404;
 import shop.mtcoding.blog.user.User;
 
 import java.util.List;
@@ -19,14 +22,24 @@ public class BoardController {
 
     @PostMapping("/board/{id}/update")
     public String update(@PathVariable Integer id, BoardRequest.UpdateDTO requestDTO) {
-        boardRepository.updateById(id, requestDTO);
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
 
+        if (sessionUser.getId() != board.getUser().getId()) {
+            throw new Exception403("게시글 수정 권한 노노");
+        }
+
+        boardRepository.updateById(id, requestDTO);
         return "redirect:/board/" + id;
     }
 
     @GetMapping("/board/{id}/update-form")
     public String updateForm(@PathVariable Integer id, HttpServletRequest request) {
         Board board = boardRepository.findById(id);
+
+        if (board == null) {
+            throw new Exception404("게시글이 없어요");
+        }
         request.setAttribute("board", board);
 
         return "board/update-form";
@@ -35,14 +48,20 @@ public class BoardController {
 
     @PostMapping("/board/{id}/delete")
     public String delete(@PathVariable Integer id) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
         Board board = boardRepository.findById(id);
+
+        if (sessionUser.getId() != board.getUser().getId()) {
+            throw new Exception403("게시글 삭제 권한 없음");
+        }
+
         boardRepository.deleteById(board.getId());
 
         return "redirect:/";
     }
 
     //완
-    @GetMapping("/" )
+    @GetMapping("/")
     public String index(HttpServletRequest request) {
         List<Board> boardList = boardRepository.findAll();
         request.setAttribute("boardList", boardList);
@@ -67,9 +86,20 @@ public class BoardController {
     //완
     @GetMapping("/board/{id}")
     public String detail(@PathVariable Integer id, HttpServletRequest request) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
         Board board = boardRepository.findByIdJoinUser(id);
+
+        boolean isBoardOwner = false;
+        if (sessionUser != null) {
+            if (sessionUser.getId() == board.getUser().getId()) {
+                isBoardOwner = true;
+            }
+        }
         request.setAttribute("board", board);
+        request.setAttribute("isBoardOwner", isBoardOwner);
 
         return "board/detail";
+
     }
+
 }
